requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Data & Finance
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.26.0

# Visualization
plotly>=5.18.0

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0

# LLM Integration (Ollama)
ollama>=0.1.0

# Scheduling
apscheduler>=3.10.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# PDF Reports
reportlab>=4.0.0

# Date utilities
python-dateutil>=2.8.0

# Hardware Detection
psutil>=5.9.0
GPUtil>=1.4.0

# Reddit API
praw>=7.7.0

# YouTube transcripts
yt-dlp>=2023.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
